+++ packages/sdk-ts/src/client/exchange/transformers/ExchangeGrpcDerivativeTransformer.ts
@@ -43,6 +43,7 @@ import {
   TradesResponse as DerivativeTradesResponse,
   PositionsResponse as DerivativePositionsResponse,
   OrderbooksResponse as DerivativeOrderbooksResponse,
+  SubaccountTradesListResponse as DerivativeSubaccountTradesListResponse,
 } from '@injectivelabs/exchange-api/injective_derivative_exchange_rpc_pb'
 import {
   BinaryOptionsMarketsResponse as BinaryOptionsMarketsResponse,
@@ -147,14 +148,20 @@ export class ExchangeGrpcDerivativeTransformer {
     return ExchangeGrpcDerivativeTransformer.grpcTradesToTrades(trades)
   }
 
+  static subaccountTradesListResponseToTrades(
+    response: DerivativeSubaccountTradesListResponse,
+  ) {
+    const trades = response.getTradesList()
+
+    return ExchangeGrpcDerivativeTransformer.grpcTradesToTrades(trades)
+  }
+
   static fundingPaymentsResponseToFundingPayments(
     response: FundingPaymentsResponse,
   ) {
     const fundingPayments = response.getPaymentsList()
 
-    return ExchangeGrpcDerivativeTransformer.grpcFundingPaymentsToFundingPayments(
-      fundingPayments,
-    )
+    return ExchangeGrpcDerivativeTransformer.grpcFundingPaymentsToFundingPayments(fundingPayments)
   }
 
   static fundingRatesResponseToFundingRates(response: FundingRatesResponse) {