+++ packages/sdk-ts/src/client/indexer/grpc/IndexerGrpcDerivativesApi.ts
@@ -153,16 +153,22 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
 
   async fetchOrders(params?: {
     marketId?: string
+    marketIds?: string[]
     orderSide?: DerivativeOrderSide
     subaccountId?: string
     pagination?: PaginationOption
   }) {
-    const { marketId, subaccountId, orderSide, pagination } = params || {}
+    const { marketId, marketIds, subaccountId, orderSide, pagination } =
+      params || {}
 
     const request = new DerivativeOrdersRequest()
 
     if (marketId) {
       request.setMarketId(marketId)
+    }
+
+    if (marketIds) {
+      request.setMarketIdsList(marketIds)
     }
 
     if (subaccountId) {
@@ -181,6 +187,10 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
       if (pagination.limit !== undefined) {
         request.setLimit(pagination.limit)
       }
+
+      if (pagination.endTime !== undefined) {
+        request.setEndTime(pagination.endTime)
+      }
     }
 
     try {
@@ -198,15 +208,20 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
 
   async fetchPositions(params?: {
     marketId?: string
+    marketIds?: string[]
     subaccountId?: string
     pagination?: PaginationOption
   }) {
-    const { marketId, subaccountId, pagination } = params || {}
+    const { marketId, marketIds, subaccountId, pagination } = params || {}
 
     const request = new DerivativePositionsRequest()
 
     if (marketId) {
       request.setMarketId(marketId)
+    }
+
+    if (marketIds) {
+      request.setMarketIdsList(marketIds)
     }
 
     if (subaccountId) {
@@ -220,6 +235,10 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
 
       if (pagination.limit !== undefined) {
         request.setLimit(pagination.limit)
+      }
+
+      if (pagination.endTime !== undefined) {
+        request.setEndTime(pagination.endTime)
       }
     }
 
@@ -244,6 +263,7 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
     subaccountId?: string
     startTime?: number
     endTime?: number
+    executionType?: TradeExecutionType
     executionSide?: TradeExecutionSide
     pagination?: PaginationOption
     marketIds?: string[]
@@ -255,6 +275,7 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
       endTime,
       direction,
       pagination,
+      executionType,
       executionSide,
       marketIds,
     } = params || {}
@@ -265,6 +286,10 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
       request.setMarketId(marketId)
     }
 
+    if (marketIds) {
+      request.setMarketIdsList(marketIds)
+    }
+
     if (subaccountId) {
       request.setSubaccountId(subaccountId)
     }
@@ -273,6 +298,10 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
       request.setMarketIdsList(marketIds)
     } else {
       request.setMarketIdsList([])
+    }
+
+    if (executionType) {
+      request.setExecutionType(executionType)
     }
 
     if (executionSide) {
@@ -299,6 +328,10 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
       if (pagination.limit !== undefined) {
         request.setLimit(pagination.limit)
       }
+
+      if (pagination.endTime !== undefined) {
+        request.setEndTime(pagination.endTime)
+      }
     }
 
     try {
@@ -316,10 +349,11 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
 
   async fetchFundingPayments(params?: {
     marketId?: string
+    marketIds?: string[]
     subaccountId?: string
     pagination?: PaginationOption
   }) {
-    const { marketId, subaccountId, pagination } = params || {}
+    const { marketId, marketIds, subaccountId, pagination } = params || {}
 
     const request = new FundingPaymentsRequest()
 
@@ -331,6 +365,10 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
       request.setSubaccountId(subaccountId)
     }
 
+    if (marketIds) {
+      request.setMarketIdsList(marketIds)
+    }
+
     if (pagination) {
       if (pagination.skip !== undefined) {
         request.setSkip(pagination.skip)
@@ -338,6 +376,10 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
 
       if (pagination.limit !== undefined) {
         request.setLimit(pagination.limit)
+      }
+
+      if (pagination.endTime !== undefined) {
+        request.setEndTime(pagination.endTime)
       }
     }
 
@@ -478,7 +520,9 @@ export class IndexerGrpcDerivativesApi extends BaseConsumer {
         typeof InjectiveDerivativeExchangeRPC.SubaccountTradesList
       >(request, InjectiveDerivativeExchangeRPC.SubaccountTradesList)
 
-      return IndexerGrpcDerivativeTransformer.tradesResponseToTrades(response)
+      return IndexerGrpcDerivativeTransformer.subaccountTradesListResponseToSubaccountTradesList(
+        response,
+      )
     } catch (e: any) {
       throw new Error(e.message)
     }