+++ packages/sdk-ts/src/client/indexer/transformers/IndexerGrpcDerivativeTransformer.ts
@@ -43,11 +43,13 @@ import {
   TradesResponse as DerivativeTradesResponse,
   PositionsResponse as DerivativePositionsResponse,
   OrderbooksResponse as DerivativeOrderbooksResponse,
+  SubaccountTradesListResponse as DerivativeSubaccountTradesListResponse,
 } from '@injectivelabs/indexer-api/injective_derivative_exchange_rpc_pb'
 import {
   BinaryOptionsMarketsResponse as BinaryOptionsMarketsResponse,
   BinaryOptionsMarketResponse as BinaryOptionsMarketResponse,
 } from '@injectivelabs/indexer-api/injective_derivative_exchange_rpc_pb'
+import { grpcPagingToPaging } from '../../../utils/pagination'
 
 const zeroPositionDelta = () => ({
   tradeDirection: TradeDirection.Buy,
@@ -131,38 +133,69 @@ export class IndexerGrpcDerivativeTransformer {
 
   static ordersResponseToOrders(response: DerivativeOrdersResponse) {
     const orders = response.getOrdersList()
+    const pagination = response.getPaging()
 
-    return IndexerGrpcDerivativeTransformer.grpcOrdersToOrders(orders)
+    return {
+      orders: IndexerGrpcDerivativeTransformer.grpcOrdersToOrders(orders),
+      pagination: grpcPagingToPaging(pagination),
+    }
   }
 
   static positionsResponseToPositions(response: DerivativePositionsResponse) {
     const positions = response.getPositionsList()
+    const pagination = response.getPaging()
 
-    return IndexerGrpcDerivativeTransformer.grpcPositionsToPositions(positions)
+    return {
+      positions:
+        IndexerGrpcDerivativeTransformer.grpcPositionsToPositions(positions),
+      pagination: grpcPagingToPaging(pagination),
+    }
   }
 
   static tradesResponseToTrades(response: DerivativeTradesResponse) {
     const trades = response.getTradesList()
+    const pagination = response.getPaging()
 
-    return IndexerGrpcDerivativeTransformer.grpcTradesToTrades(trades)
+    return {
+      trades: IndexerGrpcDerivativeTransformer.grpcTradesToTrades(trades),
+      pagination: grpcPagingToPaging(pagination),
+    }
+  }
+
+  static subaccountTradesListResponseToSubaccountTradesList(
+    response: DerivativeSubaccountTradesListResponse,
+  ) {
+    const tradesList = response.getTradesList()
+
+    return IndexerGrpcDerivativeTransformer.grpcTradesToTrades(tradesList)
   }
 
   static fundingPaymentsResponseToFundingPayments(
     response: FundingPaymentsResponse,
   ) {
     const fundingPayments = response.getPaymentsList()
+    const pagination = response.getPaging()
 
-    return IndexerGrpcDerivativeTransformer.grpcFundingPaymentsToFundingPayments(
-      fundingPayments,
-    )
+    return {
+      fundingPayments:
+        IndexerGrpcDerivativeTransformer.grpcFundingPaymentsToFundingPayments(
+          fundingPayments,
+        ),
+      pagination: grpcPagingToPaging(pagination),
+    }
   }
 
   static fundingRatesResponseToFundingRates(response: FundingRatesResponse) {
     const fundingRates = response.getFundingRatesList()
+    const pagination = response.getPaging()
 
-    return IndexerGrpcDerivativeTransformer.grpcFundingRatesToFundingRates(
-      fundingRates,
-    )
+    return {
+      fundingRates:
+        IndexerGrpcDerivativeTransformer.grpcFundingRatesToFundingRates(
+          fundingRates,
+        ),
+      pagination: grpcPagingToPaging(pagination),
+    }
   }
 
   static orderbookResponseToOrderbook(response: DerivativeOrderbookResponse) {