+++ packages/sdk-ts/src/core/authz/msgs/MsgGrant.ts
@@ -13,6 +13,7 @@ export declare namespace MsgGrant {
     grantee: string
     granter: string
     expiryInYears?: number
+    expiryInSeconds?: number
   }
 
   export interface DirectSign {
@@ -120,12 +121,12 @@ export default class MsgGrant extends MsgBase<
     return {
       '@type': '/cosmos.authz.v1beta1.MsgGrant',
       ...rest,
-    } as unknown as MsgGrant.Web3
+    }
   }
 
   private getTimestamp() {
     const { params } = this
-    const defaultExpiryYears = 5
+    const defaultExpiryYears = params.expiryInSeconds ? 0 : 5
     const dateNow = new Date()
     const expiration = new Date(
       dateNow.getFullYear() + (params.expiryInYears || defaultExpiryYears),
@@ -134,7 +135,9 @@ export default class MsgGrant extends MsgBase<
     )
 
     const timestamp = new Timestamp()
-    timestamp.setSeconds(expiration.getTime() / 1000)
+    timestamp.setSeconds(
+      expiration.getTime() / 1000 + (params.expiryInSeconds || 0),
+    )
 
     return timestamp
   }