+++ packages/sdk-ui-ts/src/client/transformers/UiAccountTransformer.ts
@@ -1,7 +1,7 @@
 import {
   SubaccountBalance,
   SubaccountTransfer,
-} from '@injectivelabs/sdk-ts/dist/client/exchange/types/account'
+} from '@injectivelabs/sdk-ts/dist/client/indexer/types/account'
 import { UiSubaccountBalance } from '../types/account'
 
 export class UiAccountTransformer {